requirements.txt
kivy==2.2.1
Pillow==9.5.0
